# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Algo Trader API
  description: Algorithmic trading platform API
  version: 3.1.9
  license: { name: Proprietary }
servers:
  - url: http://localhost:8787
    description: Local dev
  - url: https://algo-trader.workers.dev
    description: Production
security: [{ BearerAuth: [] }]
...
paths:
...
  /api/v1/co-pilot/ask:
    post:
      tags: [Co-pilot]
      summary: Ask question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties: { question: { type: string } }
      responses:
        '200':
          description: Response
          content:
            application/json:
              schema:
                type: object
                properties: { answer: { type: string }, confidence: { type: number } }
...
  /api/kyc/init:
    post:
      tags: [KYC]
      summary: Initiate KYC verification (PRO+)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId]
              properties:
                templateId:
                  type: string
                  description: Persona verification template ID
      responses:
        '200':
          description: KYC verification initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycVerification'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /api/kyc/status:
    get:
      tags: [KYC]
      summary: Get current tenant KYC status (FREE+)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: KYC status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycVerification'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /api/kyc/status/{tenantId}:
    get:
      tags: [KYC]
      summary: "Admin: get any tenant KYC status (ENTERPRISE only)"
      security:
        - BearerAuth: []
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: KYC status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycVerification'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error
  /api/kyc/webhook:
    post:
      tags: [KYC]
      summary: Persona verification webhook callback (no auth)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenantId:
                  type: string
                status:
                  type: string
                  enum: [approved, rejected, pending, expired]
                id:
                  type: string
                  description: Persona verification ID
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
                  verificationId:
                    type: string
                  status:
                    type: string
        '401':
          description: Invalid webhook signature
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  schemas:
...
    KycVerification:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        verificationId: { type: string, description: Persona verification ID }
        templateId: { type: string }
        createdAt: { type: string, format: date-time }
        verifiedAt: { type: string, format: date-time, nullable: true }
...